matrix now row-major
